pyproject.toml
[project]
name = "sg-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "tqdm>=4.67.3",
]
